is_iterator.h code moved to the corresponding header file that uses it
