--- include/common/is_iterator.h
@@ -1,35 +1,0 @@
-/**
-* @version		GrPPI v0.1
-* @copyright		Copyright (C) 2017 Universidad Carlos III de Madrid. All rights reserved.
-* @license		GNU/GPL, see LICENSE.txt
-* This program is free software: you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation, either version 3 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*
-* You have received a copy of the GNU General Public License in LICENSE.txt
-* also available in <http://www.gnu.org/licenses/gpl.html>.
-*
-* See COPYRIGHT.txt for copyright notices and details.
-*/
-
-#ifndef PPI_IS_ITERATOR
-#define PPI_IS_ITERATOR
-
-template<typename T, typename = void>
-struct is_iterator
-{
-   static constexpr bool value = false;
-};
-
-template<typename T>
-struct is_iterator<T, typename std::enable_if<!std::is_same<typename std::iterator_traits<T>::value_type, void>::value>::type>
-{
-   static constexpr bool value = true;
-};
-#endif
