+++ include/reduce.h
@@ -38,6 +38,20 @@
 	#include "include/ppi_tbb/reduce_tbb.h"
 #endif
 
+template<typename T, typename = void>
+struct is_iterator
+{
+   static constexpr bool value = false;
+};
+
+template<typename T>
+struct is_iterator<T, typename std::enable_if<!std::is_same<typename std::iterator_traits<T>::value_type, void>::value>::type>
+{
+   static constexpr bool value = true;
+};
+
+
+
 #if 0 /* START DOCUMENTATION */
 /** @addtogroup BDataPattern
  *  @{
